refactor: sub module for place service
